change dtype of DataFrame from int32 to int64

+++ taxcalc/simpletaxio.py
@@ -438,7 +438,7 @@ class SimpleTaxIO(object):
             zero_dict[names[1]] = 0
         dict_list = [zero_dict for _ in range(0, len(self._input))]
         # use dict_list to create a Pandas DataFrame and Records object
-        recsdf = pd.DataFrame(dict_list, dtype='int32')
+        recsdf = pd.DataFrame(dict_list, dtype='int64')
         recs = Records(data=recsdf, start_year=2013)
         assert recs.dim == len(self._input)
         # specify input for each tax filing unit in Records object
